Move unrelated clases from the package with resource operators (#9882)
