+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/Operator.java
@@ -6,6 +6,7 @@ package io.strimzi.operator.cluster.operator.assembly;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.strimzi.operator.common.Reconciliation;
+import io.strimzi.operator.common.config.ConfigParameter;
 import io.strimzi.operator.common.metrics.OperatorMetricsHolder;
 import io.strimzi.operator.common.model.NamespaceAndName;
 import io.vertx.core.AsyncResult;
@@ -73,7 +74,7 @@ public interface Operator {
      * @param handler       Handler called on completion.
      */
     default void reconcileThese(String trigger, Set<NamespaceAndName> desiredNames, String namespace, Handler<AsyncResult<Void>> handler) {
-        if (namespace.equals("*")) {
+        if (namespace.equals(ConfigParameter.ANY_NAMESPACE)) {
             metrics().resetResourceAndPausedResourceCounters();
         } else {
             metrics().resourceCounter(namespace).set(0);
