+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/AbstractNamespacedResourceOperator.java
@@ -20,6 +20,7 @@ import io.fabric8.kubernetes.client.informers.SharedIndexInformer;
 import io.strimzi.operator.common.Reconciliation;
 import io.strimzi.operator.common.ReconciliationLogger;
 import io.strimzi.operator.common.VertxUtil;
+import io.strimzi.operator.common.config.ConfigParameter;
 import io.strimzi.operator.common.model.Labels;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
@@ -46,12 +47,6 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
             R extends Resource<T>>
         extends AbstractResourceOperator<C, T, L, R> {
     private static final ReconciliationLogger LOGGER = ReconciliationLogger.create(AbstractNamespacedResourceOperator.class);
-
-    /**
-     * Marker for indication "all namespaces" => this is used for example when creating watches to create a cluster
-     * wide watch.
-     */
-    public final static String ANY_NAMESPACE = "*";
 
     /**
      * Constructor.
@@ -323,7 +318,7 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
      * @return  Operation with applied namespace
      */
     private FilterWatchListDeletable<T, L, R> applyNamespace(String namespace)  {
-        if (ANY_NAMESPACE.equals(namespace))  {
+        if (ConfigParameter.ANY_NAMESPACE.equals(namespace))  {
             return operation().inAnyNamespace();
         } else {
             return operation().inNamespace(namespace);
