+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/AbstractWatchableNamespacedResourceOperator.java
@@ -12,6 +12,7 @@ import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
 import io.fabric8.kubernetes.client.dsl.FilterWatchListDeletable;
 import io.fabric8.kubernetes.client.dsl.Resource;
+import io.strimzi.operator.common.config.ConfigParameter;
 import io.vertx.core.Vertx;
 
 /**
@@ -56,7 +57,7 @@ public abstract class AbstractWatchableNamespacedResourceOperator<
      * @return  A Kubernetes watch instance
      */
     public Watch watch(String namespace, Watcher<T> watcher) {
-        if (ANY_NAMESPACE.equals(namespace))    {
+        if (ConfigParameter.ANY_NAMESPACE.equals(namespace))    {
             return watchInAnyNamespace(watcher);
         } else {
             return watchInNamespace(namespace, watcher);
@@ -74,7 +75,7 @@ public abstract class AbstractWatchableNamespacedResourceOperator<
      */
     public Watch watch(String namespace, LabelSelector selector, Watcher<T> watcher) {
         FilterWatchListDeletable<T, L, R> operation
-                = ANY_NAMESPACE.equals(namespace) ? operation().inAnyNamespace() : operation().inNamespace(namespace);
+                = ConfigParameter.ANY_NAMESPACE.equals(namespace) ? operation().inAnyNamespace() : operation().inNamespace(namespace);
         if (selector != null) {
             operation = operation.withLabelSelector(selector);
         }
